Massive speed improvement
The smallest positive number that is evenly divisible by all of the
numbers from 1-2 is 2.

The smallest positive number that is evenly divisible by all of the
numbers from 1-3 is 6.

The smallest positive number that is evenly divisible by all of the
numbers from 1-4 is 12.

The smallest positive number that is evenly divisible by all of the
numbers from 1-5 is 60

60 is divisible by 12, 6 and 2. using this fact we can make this
caculation much faster by increacing our counter "n" by the result of
the previous range